SQL reservations table double fix, and reservations now save in new format

+++ wp_book_me/includes/class-wp_book_me-activator.php
@@ -102,8 +102,8 @@ class Wp_book_me_Activator {
 				startTime VARCHAR(150),
 				endTime VARCHAR(150),
 				resDate varchar(40),
-				startDoubleTime DOUBLE(2,2),
-				endDoubleTime DOUBLE(2,2)
+				startDoubleTime DOUBLE(4,2),
+				endDoubleTime DOUBLE(4,2)
 				) DEFAULT CHARACTER SET utf8";
 
 			require_once(ABSPATH . 'wp-admin/includes/upgrade.php');
